+++ CHANGELOG.md
@@ -15,14 +15,20 @@ Changelog
 - **[FIR]**: Don't use a memoizing sequence for all FirSession instance as it seems that the IDE will mutate the underlying source lists in some cases.
 - **[FIR]**: Providers can now return instances of classes nested in the same container class.
 - **[IR]**: Fix codegen error when a scoped binding in a child graph supersedes the same-typed scoped binding from a parent graph and is used in a grandchild graph's multibinding. Basically, if graph A provides `Logger` and graph `B` also provides `Logger` (overriding `A`'s), graph `C` would incorrectly try to get it from `A` instead of `B`.
+- **[metrox-viewmodel-compose]**: Pass `CreationExtras` to the `createViewModel` lamba for `assistedMetroViewModel` when using `ManualViewModelAssistedFactory`
 
 ### Changes
 
 - Test Kotlin 2.3.20-RC.
+- Test Kotlin 2.3.20-RC2.
 - Test Android Studio Panda 2
 - Test Android Studio Panda 3 canaries
 
 ### Contributors
+
+Special thanks to the following contributors for contributing to this release!
+
+- [@hvisser](https://github.com/hvisser)
 
 0.11.2
 ------